Extract separate files for plugins code
